requirements.txt
antlr4-python3-runtime==4.13.2
graphviz
PyQt5